requirements.txt
fastapi
pillow
python-multipart
numpy
uvicorn
tensorflow
requests
